requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-rest-durin==1.1.0
djangorestframework==3.14.0
durin==0.0.84
humanize==4.9.0
inputs==0.5
numpy==1.26.3
psycopg2-binary==2.9.9
pycapnp==1.3.0
pygame==2.5.2
pytz==2024.1
pyusb==1.2.1
sqlparse==0.4.4
typing_extensions==4.9.0
